More 64-bit Irix work.  This version actually compiles and runs.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@285 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_compumotor.h
@@ -65,7 +65,7 @@ MX_API mx_status_type mxd_compumotor_set_parameter( MX_MOTOR *motor );
 MX_API mx_status_type mxd_compumotor_simultaneous_start( long num_motor_records,
 						MX_RECORD **motor_record_array,
 						double *position_array,
-						int flags );
+						unsigned long flags );
 MX_API mx_status_type mxd_compumotor_get_status( MX_MOTOR *motor );
 
 MX_API mx_status_type mxd_compumotor_enable_continuous_mode(
